renderer rework wip
