+++ include/bx/framework/systems/renderer.hpp
@@ -7,9 +7,7 @@
 class Renderer : public System
 {
 public:
-	void SetPipelineOverride(const GraphicsHandle pipeline);
-
-	Vec4i GetLightIndices(const Vec3& pos);
+	/*Vec4i GetLightIndices(const Vec3& pos);
 
 	void UpdateAnimators();
 	void UpdateCameras();
@@ -20,7 +18,7 @@ public:
 	void DrawCommand(const GraphicsHandle pipeline, u32 numResourceBindings, const GraphicsHandle* pResourcesBindings, u32 numBuffers, const GraphicsHandle* pBuffers, const u64* offset, const GraphicsHandle indexBuffer, u32 count);
 	void DrawCommands();
 
-	void BindConstants(const Mat4& viewMtx, const Mat4& projMtx, const Mat4& viewProjMtx);
+	void BindConstants(const Mat4& viewMtx, const Mat4& projMtx, const Mat4& viewProjMtx);*/
 
 private:
 	void Initialize() override;
@@ -28,8 +26,4 @@ private:
 
 	void Update() override;
 	void Render() override;
-
-private:
-	class Impl;
-	Impl* m_impl = nullptr;
 };
